The scrollview of the chat has been converted into an independent UI element
Some functionality of the ChatView has been moved to the new ChatMessageViewerElement script.
