+++ Explorer/Assets/DCL/Chat/Assets/ChatMessageViewerElement.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ee73406bf16a6fe4b9d216a8a192a765
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
