+++ Explorer/Assets/DCL/Chat/ChatView.cs
@@ -1,12 +1,9 @@
 using Cysharp.Threading.Tasks;
 using DCL.Audio;
-using DCL.Clipboard;
 using DCL.Emoji;
-using DCL.Input;
 using DCL.UI;
 using MVC;
 using DG.Tweening;
-using SuperScrollView;
 using System;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
@@ -24,35 +21,14 @@ namespace DCL.Chat
     // Note: The view never changes any data (chatMessages), that's done by the controller
     public class ChatView : ViewBase, IView, IViewWithGlobalDependencies, IPointerEnterHandler, IPointerExitHandler, IDisposable
     {
-        /// <summary>
-        /// The prefab to use when instantiating a new item.
-        /// </summary>
-        private enum ChatItemPrefabIndex // It must match the list in the LoopListView.
-        {
-            ChatEntry,
-            ChatEntryOwn,
-            Padding,
-            SystemChatEntry
-        }
-
-        public delegate Color CalculateUsernameColorDelegate(ChatMessage chatMessage);
         public delegate void InputBoxFocusChangedDelegate(bool hasFocus);
         public delegate void EmojiSelectionVisibilityChangedDelegate(bool isVisible);
         public delegate void InputSubmittedDelegate(ChatChannel channel, string message, string origin);
-        public delegate void ChatMessageCreatedDelegate(int itemIndex);
 
         [Header("Settings")]
         [Tooltip("The time it takes, in seconds, for the background of the chat window to fade-in/out when hovering with the mouse.")]
         [SerializeField]
         private float BackgroundFadeTime = 0.2f;
-
-        [Tooltip("The time it takes, in seconds, for a new chat entry to fade-in.")]
-        [SerializeField]
-        private float ChatEntriesFadeTime = 3f;
-
-        [Tooltip("The time it takes, in milliseconds, without focus before the entire chat window starts fading out.")]
-        [SerializeField]
-        private int ChatEntriesWaitBeforeFading = 10000;
 
         [Tooltip("The maximum amount of character allowed in the input box.")]
         [SerializeField]
@@ -62,14 +38,6 @@ namespace DCL.Chat
         [FormerlySerializedAs("ChatBubblesToggle")]
         [SerializeField]
         private ToggleView chatBubblesToggle;
-
-        [FormerlySerializedAs("EmojiPanel")]
-        [SerializeField]
-        private EmojiPanelView emojiPanel;
-
-        [FormerlySerializedAs("EmojiSuggestionPanel")]
-        [SerializeField]
-        private EmojiSuggestionPanelView emojiSuggestionPanel;
 
         [FormerlySerializedAs("InputField")]
         [SerializeField]
@@ -83,27 +51,30 @@ namespace DCL.Chat
         [SerializeField]
         private CanvasGroup panelBackgroundCanvasGroup;
 
-        [FormerlySerializedAs("ScrollbarCanvasGroup")]
+        [FormerlySerializedAs("CloseChatButton")]
         [SerializeField]
-        private CanvasGroup scrollbarCanvasGroup;
+        private Button closeChatButton;
 
-        [FormerlySerializedAs("ChatEntriesCanvasGroup")]
         [SerializeField]
-        private CanvasGroup chatEntriesCanvasGroup;
+        private RectTransform pastePopupPosition;
 
-        [FormerlySerializedAs("LoopList")]
         [SerializeField]
-        private LoopListView2 loopList;
+        private ChatMessageViewerElement chatMessageViewer;
+
+        [Header("Emojis")]
+
+        [FormerlySerializedAs("EmojiPanel")]
+        [SerializeField]
+        private EmojiPanelView emojiPanel;
+
+        [FormerlySerializedAs("EmojiSuggestionPanel")]
+        [SerializeField]
+        private EmojiSuggestionPanelView emojiSuggestionPanel;
 
         [FormerlySerializedAs("EmojiPanelButton")]
         [SerializeField]
         private EmojiButtonView emojiPanelButton;
 
-        [FormerlySerializedAs("CloseChatButton")]
-        [SerializeField]
-        private Button closeChatButton;
-
-        [Header("Dependencies")]
         [SerializeField]
         private EmojiPanelConfigurationSO emojiPanelConfiguration;
 
@@ -118,9 +89,6 @@ namespace DCL.Chat
 
         [SerializeField]
         private EmojiSuggestionView emojiSuggestionViewPrefab;
-
-        [SerializeField]
-        private RectTransform pastePopupPosition;
 
         [Header("Audio")]
         [SerializeField]
@@ -171,11 +139,6 @@ namespace DCL.Chat
         /// </summary>
         public event Action<bool>? ChatBubbleVisibilityChanged;
 
-        /// <summary>
-        /// An external function that provides a way to calculate the color to be used to display a user name.
-        /// </summary>
-        public CalculateUsernameColorDelegate CalculateUsernameColor;
-
         private const string EMOJI_SUGGESTION_PATTERN = @":\w+";
         private const string EMOJI_TAG = "[emoji]";
         private const string ORIGIN = "chat";
@@ -189,7 +152,6 @@ namespace DCL.Chat
 
         private CancellationTokenSource emojiSearchCts;
         private CancellationTokenSource emojiPanelCts;
-        private CancellationTokenSource fadeoutCts;
         private UniTaskCompletionSource closePopupTask;
 
         private Mouse device;
@@ -198,19 +160,7 @@ namespace DCL.Chat
         private bool isInputSelected;
         private IReadOnlyDictionary<ChatChannel.ChannelId, ChatChannel>? channels;
         private ChatChannel? currentChannel;
-
-        // The latest amount of messages added to the chat that must be animated yet
-        private int entriesPendingToAnimate;
-
-        /// <summary>
-        /// Gets whether the scroll view is showing the bottom of the content, and it can't scroll down anymore.
-        /// </summary>
-        public bool IsScrollAtBottom => loopList.ScrollRect.normalizedPosition.y <= 0.0f;
-
-        /// <summary>
-        /// Gets whether the scroll view is showing the top of the content, and it can't scroll up anymore.
-        /// </summary>
-        public bool IsScrollAtTop => loopList.ScrollRect.normalizedPosition.y >= 1.0f;
+        private ChatEntryConfigurationSO chatEntryConfiguration;
 
         /// <summary>
         /// Get or sets the current content of the input box.
@@ -249,10 +199,11 @@ namespace DCL.Chat
             viewDependencies = dependencies;
         }
 
-        public void Initialize(IReadOnlyDictionary<ChatChannel.ChannelId, ChatChannel> chatChannels, ChatChannel.ChannelId defaultChannelId, bool areChatBubblesVisible)
+        public void Initialize(IReadOnlyDictionary<ChatChannel.ChannelId, ChatChannel> chatChannels, ChatChannel.ChannelId defaultChannelId, bool areChatBubblesVisible, ChatEntryConfigurationSO chatEntryConfiguration)
         {
             device = InputSystem.GetDevice<Mouse>();
             this.channels = chatChannels;
+            this.chatEntryConfiguration = chatEntryConfiguration;
 
             characterCounter.SetMaximumLength(inputField.characterLimit);
             characterCounter.gameObject.SetActive(false);
@@ -261,7 +212,8 @@ namespace DCL.Chat
             inputField.onSelect.AddListener(OnInputSelected);
             inputField.onDeselect.AddListener(OnInputDeselected);
             closeChatButton.onClick.AddListener(CloseChat);
-            loopList.InitListView(0, OnGetItemByIndex);
+            chatMessageViewer.Initialize(CalculateUsernameColor);
+            chatMessageViewer.ChatMessageOptionsButtonClicked += OnChatMessageOptionsButtonClicked;
 
             chatBubblesToggle.IsSoundEnabled = false;
             chatBubblesToggle.Toggle.isOn = areChatBubblesVisible;
@@ -286,9 +238,7 @@ namespace DCL.Chat
                 {
                     currentChannel = channels[value];
 
-                    // Replaces the chat items (it uses pools to store item instances so they will be reused)
-                    loopList.SetListItemCount(0);
-                    loopList.SetListItemCount(currentChannel.Messages.Count);
+                    chatMessageViewer.SetData(currentChannel.Messages);
                 }
             }
         }
@@ -300,10 +250,7 @@ namespace DCL.Chat
         public void ToggleChat(bool show)
         {
             panelBackgroundCanvasGroup.gameObject.SetActive(show);
-            loopList.gameObject.SetActive(show);
-
-            if (!show) // Note: This is necessary to avoid items animating when re-opening the chat window
-                entriesPendingToAnimate = 0;
+            chatMessageViewer.SetVisibility(show);
         }
 
         /// <summary>
@@ -322,51 +269,6 @@ namespace DCL.Chat
         public void EnableInputBoxSubmissions()
         {
             inputField.onSubmit.AddListener(OnSubmit);
-        }
-
-        /// <summary>
-        /// Moves the chat so it shows the last created message.
-        /// </summary>
-        public void ShowLastMessage()
-        {
-            loopList.MovePanelToItemIndex(0, 0);
-        }
-
-        /// <summary>
-        /// Makes sure the chat window is showing all the messages stored in the data.
-        /// </summary>
-        public void RefreshMessages()
-        {
-            ResetChatEntriesFadeout();
-            loopList.SetListItemCount(currentChannel.Messages.Count);
-
-            entriesPendingToAnimate = currentChannel.Messages.Count - loopList.ItemTotalCount;
-
-            if (entriesPendingToAnimate < 0)
-                entriesPendingToAnimate = 0;
-
-            loopList.SetListItemCount(currentChannel.Messages.Count);
-            ShowLastMessage();
-
-// DISABLED UNTIL UNREAD MESSAGES FEATURE IS MERGED
-
-            // Scroll view adjustment
-//            if (IsScrollAtBottom)
-//            {
-//                loopList.MovePanelToItemIndex(0, 0);
-//            }
-//            else
-//            {
-//                loopList.RefreshAllShownItem();
-
-                // When the scroll view is not at the bottom, chat messages should not move if a new message is added
-                // An offset has to be applied to the scroll view in order to prevent messages from moving
-//                LoopListViewItem2 addedItem = loopList.GetShownItemByIndex(1);
-//                float offsetToPreventScrollViewMovement = -addedItem.ItemSize - addedItem.Padding;
-//                loopList.MovePanelByOffset(offsetToPreventScrollViewMovement);
-
-                // Known issue: When the scroll view is at the top, the scroll view moves a bit downwards
-//            }
         }
 
         /// <summary>
@@ -400,6 +302,14 @@ namespace DCL.Chat
         public void SubmitInput()
         {
             inputField.OnSubmit(new BaseEventData(null));
+        }
+
+        /// <summary>
+        /// Makes sure the chat window is showing all the messages stored in the data for the current channel.
+        /// </summary>
+        public void RefreshMessages()
+        {
+            chatMessageViewer.RefreshMessages();
         }
 
         /// <summary>
@@ -449,16 +359,15 @@ namespace DCL.Chat
         {
             PointerEnter?.Invoke();
             panelBackgroundCanvasGroup.DOFade(1, BackgroundFadeTime);
-            scrollbarCanvasGroup.DOFade(1, BackgroundFadeTime);
-            StopChatEntriesFadeout();
+            chatMessageViewer.SetScrollbarVisibility(true, BackgroundFadeTime);
+            chatMessageViewer.StopChatEntriesFadeout();
         }
 
         public void OnPointerExit(PointerEventData eventData)
         {
             PointerExit?.Invoke();
-            panelBackgroundCanvasGroup.DOFade(0, BackgroundFadeTime);
-            scrollbarCanvasGroup.DOFade(0, BackgroundFadeTime);
-            StartChatEntriesFadeout();
+            chatMessageViewer.SetScrollbarVisibility(false, BackgroundFadeTime);
+            chatMessageViewer.StartChatEntriesFadeout();
         }
 
         public void Dispose()
@@ -472,7 +381,6 @@ namespace DCL.Chat
             if (emojiSuggestionPanelController != null)
                 emojiSuggestionPanelController.EmojiSelected -= AddEmojiFromSuggestion;
 
-            fadeoutCts.SafeCancelAndDispose();
             emojiPanelCts.SafeCancelAndDispose();
             emojiSearchCts.SafeCancelAndDispose();
 
@@ -482,90 +390,6 @@ namespace DCL.Chat
         private void Start()
         {
             panelBackgroundCanvasGroup.alpha = 0;
-            scrollbarCanvasGroup.alpha = 0;
-        }
-
-        // Called by the LoopListView when the number of items change, it uses out data (chatMessages)
-        // to customize a new instance of the ChatEntryView (it uses pools internally).
-        private LoopListViewItem2? OnGetItemByIndex(LoopListView2 listView, int index)
-        {
-            if (index < 0 || index >= currentChannel.Messages.Count)
-                return null;
-
-            ChatMessage itemData = currentChannel.Messages[index];
-            LoopListViewItem2 item;
-
-            if (itemData.IsPaddingElement)
-                item = listView.NewListViewItem(listView.ItemPrefabDataList[(int)ChatItemPrefabIndex.Padding].mItemPrefab.name);
-            else
-            {
-                item = listView.NewListViewItem(itemData.SystemMessage ? listView.ItemPrefabDataList[(int)ChatItemPrefabIndex.SystemChatEntry].mItemPrefab.name :
-                    itemData.SentByOwnUser ? listView.ItemPrefabDataList[(int)ChatItemPrefabIndex.ChatEntryOwn].mItemPrefab.name
-                                            : listView.ItemPrefabDataList[(int)ChatItemPrefabIndex.ChatEntry].mItemPrefab.name);
-
-                ChatEntryView itemScript = item!.GetComponent<ChatEntryView>()!;
-                SetItemData(index, itemData, itemScript);
-
-                Button? messageOptionsButton = itemScript.messageBubbleElement.messageOptionsButton;
-                messageOptionsButton?.onClick.RemoveAllListeners();
-
-                messageOptionsButton?.onClick.AddListener(() =>
-                    OnChatMessageOptionsButtonClicked(itemData.Message, itemScript));
-            }
-
-            return item;
-        }
-
-        private void SetItemData(int index, ChatMessage itemData, ChatEntryView itemView)
-        {
-            //temporary approach to extract the username without the walledId, will be refactored
-            //once we have the proper integration of the profile retrieval
-            Color playerNameColor = CalculateUsernameColor(itemData);
-
-            itemView.usernameElement.userName.color = playerNameColor;
-
-            if (!itemData.SystemMessage)
-            {
-                itemView.ProfileBackground!.color = playerNameColor;
-                playerNameColor.r += 0.3f;
-                playerNameColor.g += 0.3f;
-                playerNameColor.b += 0.3f;
-                itemView.ProfileOutline!.color = playerNameColor;
-            }
-
-            itemView.SetItemData(itemData);
-
-            // Views that correspond to new added items have to be animated
-            if (index - 1 < entriesPendingToAnimate) // Note: -1 because the first real message starts at 1, which is the latest messaged added
-                itemView.AnimateChatEntry();
-        }
-
-        private void StopChatEntriesFadeout()
-        {
-            fadeoutCts.SafeCancelAndDispose();
-            chatEntriesCanvasGroup.alpha = 1;
-        }
-
-        private void StartChatEntriesFadeout()
-        {
-            fadeoutCts.SafeCancelAndDispose();
-            fadeoutCts = new CancellationTokenSource();
-
-            AwaitAndFadeChatEntriesAsync(fadeoutCts.Token).Forget();
-        }
-
-        private void ResetChatEntriesFadeout()
-        {
-            StopChatEntriesFadeout();
-            StartChatEntriesFadeout();
-        }
-
-        private async UniTaskVoid AwaitAndFadeChatEntriesAsync(CancellationToken ct)
-        {
-            fadeoutCts.Token.ThrowIfCancellationRequested();
-            chatEntriesCanvasGroup.alpha = 1;
-            await UniTask.Delay(ChatEntriesWaitBeforeFading, cancellationToken: ct);
-            await chatEntriesCanvasGroup.DOFade(0.4f, ChatEntriesFadeTime).ToUniTask(cancellationToken: ct);
         }
 
         private void CloseChat()
@@ -604,7 +428,7 @@ namespace DCL.Chat
             inputField.stringPosition += textToInsert.Length;
             inputField.ActivateInputField();
             characterCounter.SetCharacterCount(inputField.text.Length);
-            StopChatEntriesFadeout();
+            chatMessageViewer.StopChatEntriesFadeout();
         }
 
         public override UniTask HideAsync(CancellationToken ct, bool isInstant = false)
@@ -618,7 +442,7 @@ namespace DCL.Chat
             isInputSelected = false;
             emojiPanelButton.SetColor(false);
             characterCounter.gameObject.SetActive(false);
-            StartChatEntriesFadeout();
+            chatMessageViewer.StartChatEntriesFadeout();
             InputBoxFocusChanged?.Invoke(false);
         }
 
@@ -628,7 +452,7 @@ namespace DCL.Chat
             {
                 isChatClosed = false;
                 ToggleChat(true);
-                ShowLastMessage();
+                chatMessageViewer.ShowLastMessage();
             }
 
             UIAudioEventsBus.Instance.SendPlayAudioEvent(enterInputAudio);
@@ -638,7 +462,7 @@ namespace DCL.Chat
             isInputSelected = true;
             emojiPanelButton.SetColor(true);
             characterCounter.gameObject.SetActive(true);
-            StopChatEntriesFadeout();
+            chatMessageViewer.StopChatEntriesFadeout();
             InputBoxFocusChanged?.Invoke(true);
         }
 
@@ -648,7 +472,7 @@ namespace DCL.Chat
             UIAudioEventsBus.Instance.SendPlayAudioEvent(chatInputTextAudio);
             closePopupTask.TrySetResult();
             characterCounter.SetCharacterCount(inputText.Length);
-            StopChatEntriesFadeout();
+            chatMessageViewer.StopChatEntriesFadeout();
         }
 
         private void OnSubmit(string _)
@@ -720,6 +544,9 @@ namespace DCL.Chat
 
         private bool IsWithinCharacterLimit() =>
             inputField.text.Length < inputField.characterLimit;
+
+        private Color CalculateUsernameColor(ChatMessage chatMessage) =>
+            chatEntryConfiguration.GetNameColor(chatMessage.SenderValidatedName);
 
         #region Emojis
